+++ dsmrreader/__init__.py
@@ -1,5 +1,5 @@
 from django.utils.version import get_version
 
-VERSION = (4, 0, 0, 'rc', 1)
+VERSION = (4, 0, 0, 'rc', 2)
 
 __version__ = get_version(VERSION)